project1/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.5.0.5
celery==4.1.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-reversion==3.0.5
django-simple-captcha==0.5.12
django-tinymce==2.4.0
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.18.2
httplib2==0.9.2
idna==2.8
importlib-metadata==1.1.0
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
kombu==4.2.0
MarkupPy==1.14
more-itertools==8.0.0
mutagen==1.43.0
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
py3Fdfs==2.2.0
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.2
redis==2.10.6
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
vine==1.3.0
Whoosh==2.5.0
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
